requirements.txt
azure-functions
Pillow
requests